+++ Futures/Result.swift
@@ -1,5 +1,3 @@
-import Foundation
-
 /// Typical Result enumeration (aka Either)
 public enum Result<T> {
   /// The result contains a Success value
@@ -10,4 +8,20 @@ public enum Result<T> {
   
   /// The result was cancelled
   case Cancelled
+  
+  //TODO: Expose in a later version
+  func map<U>(handler: T -> U) -> Future<U> {
+    let mapped = Promise<U>()
+    
+    switch self {
+    case .Success(let value):
+      mapped.succeed(handler(value))
+    case .Error(let error):
+      mapped.fail(error)
+    case .Cancelled:
+      mapped.cancel()
+    }
+    
+    return mapped.future
+  }
 }